Reuse existing lxc for new installations

+++ server_agent_provision/create_k3s_server_agent_proxmox.bash
@@ -9,6 +9,7 @@ SCRIPT_DIR="${SCRIPT_DIR%/*}"
 # Configuration variables
 # 
 DEBUG=${DEBUG:-false}
+USE_EXISTING_VMID=${0:-USE_EXISTING_VMID}
 
 SECRETS_FILE="secret"
 POOL_VMID_STARTS_AT=3000 # _VMID to start looking for free IDS in Proxmox
@@ -121,46 +122,56 @@ _pct_exec_file() {
 # START THE PROVISIONING
 # -------------------------------------------
 
-# Defining machines in the cluster
-VMID_SERVER=${POOL_VMID_STARTS_AT}
-while pct config ${VMID_SERVER} > /dev/null 2> /dev/null; do
-  VMID_SERVER=$(( VMID_SERVER + 1 ))
-done
-VMID_AGENT=$(( VMID_SERVER +1 ))
-while pct config ${VMID_AGENT} > /dev/null 2> /dev/null; do
-  VMID_AGENT=$(( VMID_AGENT + 1 ))
-done
+if [[ ${USE_EXISTING_VMID} -gt 0 ]]; then
+  VMID_SERVER=${USE_EXISTING_VMID}
+  VMID_AGENT=$(( USE_EXISTING_VMID + 1 ))
+else
+  # Defining machines in the cluster
+  VMID_SERVER=${POOL_VMID_STARTS_AT}
+  while pct config ${VMID_SERVER} > /dev/null 2> /dev/null; do
+    VMID_SERVER=$(( VMID_SERVER + 1 ))
+  done
+  VMID_AGENT=$(( VMID_SERVER +1 ))
+  while pct config ${VMID_AGENT} > /dev/null 2> /dev/null; do
+    VMID_AGENT=$(( VMID_AGENT + 1 ))
+  done
+fi
 CLUSTER_INSTANCES=()
 CLUSTER_INSTANCES[VMID_SERVER]="cluster-k3s-server-${VMID_SERVER}"
 CLUSTER_INSTANCES[VMID_AGENT]="cluster-k3s-agent-${VMID_SERVER}"
 
-# Base installation for all the instances
-for VMID in "${!CLUSTER_INSTANCES[@]}"; do
-  HOSTNAME=${CLUSTER_INSTANCES[VMID]}
-  echo "[ --- ] Creating instance ${HOSTNAME} with ID ${VMID}"
-  echo "[ RUN ] Building the PVE instance"
-  _pct_create "${VMID}" "${HOSTNAME}"
-  echo "[ RUN ] Starting the PVE instance"
-  _pct_start "${VMID}"
-  echo "[ RUN ] Base packages installation"
-  _pct_exec "${VMID}" "sed -i -e 's:# en_US.UTF-8 UTF-8:en_US.UTF-8 UTF-8:' /etc/locale.gen"
-  _pct_exec "${VMID}" "locale-gen > /dev/null 2> /dev/null"
-  _pct_exec "${VMID}" "apt-get -qq update"
-  _pct_exec "${VMID}" "apt-get install -qq -o=Dpkg::User-Pty=0 -y ${BASE_APT_PACKAGES} > /dev/null"
-  echo "[ RUN ] Prepare for the k3s installation"
-  _pct_exec_file "${VMID}" "prepare_lxc_for_k3s.bash"
+if [[ ${USE_EXISTING_VMID} -eq 0 ]]; then
+  # Base installation for all the instances
+  for VMID in "${!CLUSTER_INSTANCES[@]}"; do
+    HOSTNAME=${CLUSTER_INSTANCES[VMID]}
+    echo "[ --- ] Creating instance ${HOSTNAME} with ID ${VMID}"
+    echo "[ RUN ] Building the PVE instance"
+    _pct_create "${VMID}" "${HOSTNAME}"
+    echo "[ RUN ] Starting the PVE instance"
+    _pct_start "${VMID}"
+    echo "[ RUN ] Base packages installation"
+    _pct_exec "${VMID}" "sed -i -e 's:# en_US.UTF-8 UTF-8:en_US.UTF-8 UTF-8:' /etc/locale.gen"
+    _pct_exec "${VMID}" "locale-gen > /dev/null 2> /dev/null"
+    _pct_exec "${VMID}" "apt-get -qq update"
+    _pct_exec "${VMID}" "apt-get install -qq -o=Dpkg::User-Pty=0 -y ${BASE_APT_PACKAGES} > /dev/null"
+    echo "[ RUN ] Prepare for the k3s installation"
+    _pct_exec_file "${VMID}" "prepare_lxc_for_k3s.bash"
+    echo "[ --- ]"
+    echo
+  done
+  # Server installation
   echo "[ --- ]"
-  echo
-done
+  echo "[ SERVER ] Install the k3s server"
+  _pct_exec_file ${VMID_SERVER} "install_k3s_server.bash"
+  echo "[ SERVER ] Install the helm package manager"
+  _pct_exec_file ${VMID_SERVER} "install_helm.bash"
+  echo "[ TEST ] Check server installation"
+  _pct_exec ${VMID_SERVER} "/usr/local/bin/kubectl get nodes > /dev/null 2>/dev/null"
+else
+  echo "[ --- ]"
+  echo "[ SERVER ] Reusing server ${VMID_SERVER} and agent ${VMID_AGENT}"
+fi
 
-# Server installation
-echo "[ --- ]"
-echo "[ SERVER ] Install the k3s server"
-_pct_exec_file ${VMID_SERVER} "install_k3s_server.bash"
-echo "[ SERVER ] Install the helm package manager"
-_pct_exec_file ${VMID_SERVER} "install_helm.bash"
-echo "[ TEST ] Check server installation"
-_pct_exec ${VMID_SERVER} "/usr/local/bin/kubectl get nodes > /dev/null 2>/dev/null"
 echo "[ TEST ] Check helm installation"
 _pct_exec ${VMID_SERVER} "/usr/local/bin/helm version > /dev/null"
 echo "[ SERVER ] Install argo-cd"
